bin/requirements.txt
args==0.1.0
asn1crypto==0.24.0
beautifulsoup4==4.6.0
bitstring==3.1.5
bs4==0.0.1
build==1.0.2
cached-property==1.4.2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
clint==0.5.1
configparser==3.5.0
cryptography==2.2.2
docopt==0.6.2
entrypoints==0.2.3
fonttools==3.26.0
future==0.16.0
google-api-python-client==1.6.7
googletrans==2.3.0
gTTS==2.0.1
gTTS-token==1.1.1
httplib2==0.11.3
idna==2.6
jeepney==0.3.1
keyring==12.2.1
keyrings.alt==3.1
numpy==1.14.3
oauth2client==4.1.2
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycparser==2.18
pydub==0.22.1
pygame==1.9.4
python-dateutil==2.7.3
pytz==2018.4
pywin32-ctypes==0.1.2
requests==2.20.0
rsa==3.4.2
SecretStorage==3.0.1
six==1.11.0
speech-recognition==0.1
SpeechRecognition==3.8.1
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
